starting refactoring to add instances
