+++ spython/image/cmd/__init__.py
@@ -1,4 +1,3 @@
-
 # Copyright (C) 2018 The Board of Trustees of the Leland Stanford Junior
 # University.
 # Copyright (C) 2017-2018 Vanessa Sochat.
@@ -17,8 +16,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .generate import RobotNamer
-
 def generate_image_commands():
     ''' The Image client holds the Singularity image command group, mainly
         deprecated commands (image.import) and additional command helpers
@@ -26,14 +23,22 @@ def generate_image_commands():
 
         The levels of verbosity (debug and quiet) are passed from the main
         client via the environment variable MESSAGELEVEL.
+
+        These commands are added to Client.image under main/__init__.py to 
+        expose subcommands:
+
+            Client.image.export
+            Client.image.imprt
+            Client.image.decompress
+            Client.image.create
+
     '''
 
-    class ImageClient:
+    class ImageClient(object):
         group = "image"
 
     from spython.main.base.logger import println
-    from spython.main.base.command import ( init_command, run_command )
-
+    from spython.main.base.commands import ( init_command, run_command )
     from .utils import ( compress, decompress )
     from .create import create
     from .importcmd import importcmd
@@ -44,7 +49,6 @@ def generate_image_commands():
     ImageClient.export = export
     ImageClient.decompress = decompress
     ImageClient.compress = compress
-
     ImageClient.println = println
     ImageClient.init_command = init_command
     ImageClient.run_command = run_command
