+++ spython/main/base/__init__.py
@@ -30,6 +30,7 @@ from .command import ( generate_bind_list, init_command, run_command )
 from .flags import parse_verbosity
 from .sutils import ( get_uri, load, setenv, get_filename )
 from .logger import ( println,  init_level )
+from .generate import RobotNamer
 
 class Client:
 
@@ -72,6 +73,7 @@ class Client:
             sys.exit(1)
 
 
+
 # Image Utils
 Client.load = load
 Client._get_filename = get_filename
@@ -87,3 +89,4 @@ Client._run_command = run_command
 Client._parse_verbosity = parse_verbosity
 Client._println = println
 Client._init_level = init_level
+Client.RobotNamer = RobotNamer()
